Test: post refactor, change test into one file with integrationTest state.
